Minor changes done.
